add CodingStrategyOptions protocol, allow BSONEncoder/BSONDecoder to be configured via an options struct
